requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
brotli==1.2.0
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
dj-database-url==3.1.2
Django==5.2.11
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
hiredis==3.3.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
newrelic-telemetry-sdk==0.9.0
openpyxl==3.1.5
packaging==25.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybreaker==1.4.1
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
git+https://github.com/OtomashenGit/komp-tools-box.git@v0.5.0#egg=tools-box
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
zope.interface==8.2
pydantic==2.12.5
pydantic_core==2.41.5

